Share number might be larger than long
Total share number might overflow long max value, even if the database
number which is typically divided by 1000 doesn't. This patch changes
the database column "value" to be BIGINT, which might be overkill and we
may decide to go back to INT. At the same time does the internal
computations with long long and import with setInt64.

Also return 0 for DY when 7.08.04.01 does not have any value as is the
case in financial companies for example.

Fixes #5